requirements.txt
firebolt-sdk
matplotlib
pandas